[FLINK-37933][table] Move TransformationScan/SinkProvider out of planner
Although TransformationScan/SinkProvider is an expert API, it's buried inside planner and not easily usable. Move up to table-common to make it more accessible.
